refactor: share html formatting helpers
